refactor:숫자로만 이루어진 로또번호인지 체크로직을 LottoNumber로 옮김
